+++ R/AllClasses.R
@@ -106,11 +106,11 @@ NULL
 #' TimeSeries
 #'
 #' An S4 class to represent time series.
-#' @slot .Data A \eqn{n x m x p}{n \times m \times p} `numeric` [`array`]
+#' @slot .Data A \eqn{n \times m \times p} `numeric` [`array`]
 #'  giving the observed time-series values.
 #' @slot .Time A length-\eqn{n} [`RataDie-class`] object.
 #' @details
-#'  A time series object is an \eqn{n x m x p}{n \times m \times p} array, with
+#'  A time series object is an \eqn{n \times m \times p} array, with
 #'  \eqn{n} being the number of observations, \eqn{m} being the number of series
 #'  and with the \eqn{p} columns of the third dimension containing extra
 #'  variables for each series.