+++ R/aion-package.R
@@ -1,6 +1,6 @@
 #' @details
 #'  \tabular{ll}{
-#'   **Version** \tab 1.6.0 \cr
+#'   **Version** \tab 1.7.0 \cr
 #'   **License** \tab GPL-3 \cr
 #'   **CRAN DOI** \tab \doi{10.32614/cran.package.aion} \cr
 #'   **Zenodo DOI** \tab \doi{10.5281/zenodo.8032278} \cr