+++ R/coerce.R
@@ -16,22 +16,22 @@ setMethod(
 #' @export
 #' @method as.data.frame TimeSeries
 as.data.frame.TimeSeries <- function(x, ..., calendar = NULL) {
+  ## Fix dimnames (if needed)
+  x <- fix_dimnames(x)
   ## Build a long data frame
-  z <- as.data.frame.table(x, base = list("T", "S", LETTERS))
-
-  ## Add sampling times
-  z[[1]] <- time(x, calendar = calendar)
-
-  ## Fix colnames
-  colnames(z) <- c("time", "series", "variable", "value")
-
-  z
+  data.frame(
+    time = time(x, calendar = calendar)[as.vector(slice.index(x, MARGIN = 1L))],
+    series = dimnames(x)[[2L]][as.vector(slice.index(x, MARGIN = 2L))],
+    variable = dimnames(x)[[3L]][as.vector(slice.index(x, MARGIN = 3L))],
+    value = as.vector(x)
+  )
 }
 
 #' @export
 #' @describeIn as.data.frame Returns a long [`data.frame`] with the following columns:
 #'  \describe{
-#'   \item{`time`}{The (decimal) years at which the time series was sampled.}
+#'   \item{`time`}{The sampling times. If `calendar` is not `NULL`, it is
+#'   expressed in decimal years; otherwise, it is expressed in *rata die*.}
 #'   \item{`series`}{The name of the time series.}
 #'   \item{`variable`}{The name of the variables.}
 #'   \item{`value`}{The observed value.}
@@ -54,8 +54,11 @@ as.data.frame.TimeIntervals <- function(x, ..., calendar = NULL) {
 #' @describeIn as.data.frame Returns a [`data.frame`] with the following columns:
 #'  \describe{
 #'   \item{`label`}{The name of the intervals.}
-#'   \item{`start`}{The start time of the intervals, in (decimal) years.}
-#'   \item{`end`}{The end time of the intervals, in (decimal) years.}
+#'   \item{`start`}{The start time of the intervals. If `calendar` is not
+#'   `NULL`, it is expressed in decimal years; otherwise, it is expressed in
+#'   *rata die*.}
+#'   \item{`end`}{The end time of the intervals. If `calendar` is not `NULL`, it
+#'   is expressed in decimal years; otherwise, it is expressed in *rata die*.}
 #'  }
 #' @aliases as.data.frame,TimeIntervals-method
 setMethod("as.data.frame", "TimeIntervals", as.data.frame.TimeIntervals)